Changing implementation for to_string as well as how the size is calculated
The size function needed for push_back function was iterating through the string each time it was needed, and for each coordinates, the calculation was performed. It was taking too long to compute during the complie time, and was going through too many iterations. Changed it so that the size is a class variable that is incremented/decremented whenever appropriate, and the size() function is reduced to O(1). Also realized I was not flipping the characters in to_string, so it was giving 83 for 38 as an input.